+++ smos-docs-site/content/pages/changelog.markdown
@@ -8,8 +8,10 @@ description: The changelog for all of the Smos tools and libraries
 ## Changed
 
 - `smos-github`: There is now a new column that suggests you to have another look at the issue/pr because there has been a recent enough update.
-- `smos-github`: Done entries are no longer mentioned in the `smos-github list` report.
+- `smos-github`: Done entries are no longer mentioned in the `list` report.
+- `smos-github` Show the state of the entries in the `list` report as well.
 - `smos`: A keybinding to open the `url` property in smos: `<space>ou` by default.
+- `smos-query` always show "no todo state" as `(none)` instead of the empty string.
 
 # <a name="2021-04-09">[2021-04-09 Release](#2021-04-09)